Day 11 - star 2 solved!!! :D GCD

+++ day11/MITM/Monkey.cs
@@ -3,7 +3,7 @@ namespace AOC2022.Day11.MITM;
 public class Monkey
 {
     public int MonkeyNumber { get; set; }
-    public List<long> ItemsHeld { get; set; }
+    public List<double> ItemsHeld { get; set; }
     public string OperationName { get; set; }
     public string OperationValueType { get; set; }
     public long OperationValue { get; set; }
@@ -15,7 +15,7 @@ public class Monkey
     public Monkey()
     {
         MonkeyNumber = 0;
-        ItemsHeld = new List<long>();
+        ItemsHeld = new List<double>();
         OperationName = String.Empty;
         OperationValueType = String.Empty;
         OperationValue = 0;
